refactor: moved unique gift properties into separate package
