+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/gifts/unique/UniqueGiftBackdrop.kt
@@ -1,19 +1,18 @@
-package dev.inmo.tgbotapi.types.gifts
+package dev.inmo.tgbotapi.types.gifts.unique
 
-import dev.inmo.tgbotapi.types.files.Sticker
+import dev.inmo.tgbotapi.types.colorsField
 import dev.inmo.tgbotapi.types.nameField
 import dev.inmo.tgbotapi.types.rarityPerMilleField
-import dev.inmo.tgbotapi.types.stickerField
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 
 @Serializable
-data class UniqueGiftSymbol(
+data class UniqueGiftBackdrop(
     @SerialName(nameField)
     val name: String,
-    @SerialName(stickerField)
-    val sticker: Sticker,
+    @SerialName(colorsField)
+    val colors: UniqueGiftBackdropColors,
     @SerialName(rarityPerMilleField)
     val rarityPerMille: Int
 )
