+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/gifts/unique/UniqueGiftModel.kt
@@ -1,4 +1,4 @@
-package dev.inmo.tgbotapi.types.gifts
+package dev.inmo.tgbotapi.types.gifts.unique
 
 import dev.inmo.tgbotapi.types.files.Sticker
 import dev.inmo.tgbotapi.types.nameField
